ACQ-317
Adds ability for phone searches and some basic validation. Adds patient list modal and a way to normalize the phone so it's easily searchable in the database. Modifies the style.less so that it's easier to update the ruleset as well. Updates the locale file to account for the placeholder and modal title. Copies over the files used for common usage to the common folder so that files can start to reference that. Adds reducer, action, saga, and selector code to allow for searches to be displayed and working.